#3: Move common mock utils to MockUtils.
